Refactoring task 18
