+++ app/src/main/java/com/example/administrator/readaloud/database/DBHelper.java
@@ -1,0 +1,11 @@
+package com.example.administrator.readaloud.database;
+
+import android.database.sqlite.SQLiteOpenHelper;
+
+/**
+ * Created by Administrator on 05.02.2018.
+ */
+
+public class DBHelper {
+
+}
