+++ app/src/main/java/com/example/administrator/readaloud/ui/BaseToolbar.java
@@ -1,4 +1,4 @@
-package com.example.administrator.readaloud;
+package com.example.administrator.readaloud.ui;
 
 import android.support.v4.widget.DrawerLayout;
 import android.support.v7.app.ActionBar;
