+++ app/src/main/java/com/example/administrator/readaloud/ui/MainActivity.java
@@ -1,8 +1,9 @@
-package com.example.administrator.readaloud;
+package com.example.administrator.readaloud.ui;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
+import com.example.administrator.readaloud.R;
 import com.example.administrator.readaloud.ui.welcome.WelcomeFragment;
 
 public class MainActivity extends AppCompatActivity {
