+++ app/src/main/res/layout/toolbar_main.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <android.support.v7.widget.Toolbar xmlns:android="http://schemas.android.com/apk/res/android"
-    android:id="@+id/toolbarMain"
+    android:id="@+id/toolbar_main"
     android:layout_width="match_parent"
     android:layout_height="64dp"
     android:background="@color/primaryColor"
